requirements.txt
scikit-learn
feast
pandas
requests
python-dotenv
pyarrow
joblib
numpy
matplotlib
shap
streamlit
fastapi
uvicorn
jinja2